requirements.txt
selene~=2.0.0rc9
pytest~=8.3.4
python-dotenv~=1.0.1
selenium~=4.27.1